requirements.txt
tinydb>=4.7.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
pyttsx3>=2.90
pyautogui>=0.9.54
playwright>=1.40.0
vosk>=0.3.45
pyaudio>=0.2.14
google-generativeai>=0.3.0
python-telegram-bot>=20.6
pywin32>=306